Improved fuzzy search for organisations
The initial search now uses QuerySets and filters, meaning only one query
is actually executed. Also, it now uses icontains instead of iregex, since
it was only doing a contains check anyway. Finally, if a school with no
admin exists, this school is not returned (as no email can be sent to its
admin anyway).